feat: 1.Update: autoconfigure add ResourceCleaner;2.rename classpath reader to resource reader
