+++ popup-bom/popup-parent/popup-starter/popup-starter-autoconfigure/pom.xml
@@ -35,5 +35,10 @@
             <artifactId>popup-starter-validator</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>com.photowey</groupId>
+            <artifactId>popup-cloud-uaa-auth</artifactId>
+        </dependency>
+
     </dependencies>
 </project>
